+++ build.rs
@@ -22,11 +22,21 @@ fn main() {
     println!("cargo:rerun-if-changed=src/ffi.rs");
 
     // emit linking information AFTER compiling the bridge
-    pkg_config::Config::new()
-        .atleast_version("6")
-        .statik(cfg!(feature = "static"))
-        .probe("lhapdf")
-        .unwrap();
+    if cfg!(feature = "static") {
+        // static linking is broken in pkg-config:
+        // https://github.com/rust-lang/pkg-config-rs/issues/102, therefore we do it manually here
+        for link_path in lhapdf.link_paths {
+            println!("cargo:rustc-link-search={}", link_path.to_str().unwrap());
+        }
+        for lib in lhapdf.libs {
+            println!("cargo:rustc-link-lib=static={}", lib);
+        }
+    } else {
+        pkg_config::Config::new()
+            .atleast_version("6")
+            .probe("lhapdf")
+            .unwrap();
+    }
 }
 
 #[cfg(feature = "docs-only")]